update to 3.2.1

+++ src/Crosscutting/Infra.Crosscutting/Collections/IPagedList.cs
@@ -5,5 +5,5 @@ namespace Goal.Infra.Crosscutting.Collections;
 public interface IPagedList : IEnumerable
 {
     int PageLength { get; }
-    int TotalCount { get; }
+    long TotalCount { get; }
 }
